added the not found apis

+++ app/Http/Controllers/UserController.php
@@ -77,7 +77,7 @@ class UserController extends Controller
             if (count($user) > 0) {
                 $token = Str::random(40);
                 $domain = URL::to('/');
-                $url = $domain . '/reset-password?token=' . $token;
+                $url = $domain . '/resetpass?token=' . $token;
 
                 $data['url'] = $url;
                 $data['email'] = $request->email;
@@ -123,35 +123,72 @@ class UserController extends Controller
         }
 
     }
-    
+
+
+    // public function resetpass(Request $request)
+    // {
+    //     $resetd = password_reset::where('token', $request->token)->get();
+
+    //     if (isset($request->token) && count($resetd) > 0) {
+
+    //         $user = User::where('email', $resetd[0]['email'])->get();
+    //         return view('reset-password', compact('user'));
+
+    //     } else {
+    //         return 'No';
+    //     }
+    // }
+
+    // public function updatepass(Request $request)
+    // {
+    //     $request->validate([
+    //         'password' => 'required|string|min:6|confirmed',
+    //         'confirm_pass' => 'required|string|min:6|same:password',
+    //     ]);
+
+    //     $user = User::find($request->id);
+    //     $user->password = Hash::make($request->password);
+    //     $user->save();
+
+    //     password_reset::where('email', $user->email)->delete();
+
+    //     return "<h1>Your Password Reset was Successful!</h1>";
+
+    // }
 
     public function resetpass(Request $request)
     {
-        $resetd = password_reset::where('token', $request->token)->get();
+        if (isset($request->token)) {
+            $resetData = Password_reset::where('token', $request->token)->first();
 
-        if (isset($request->token) && count($resetd) > 0) {
-
-            $user = User::where('email', $resetd[0]['email'])->get();
-            return view('reset-password', compact('user'));
-
+            if ($resetData) {
+                $user = User::where('email', $resetData->email)->first();
+                // dd($user['email']);
+                return view('reset_password', compact('user'));
+            } else {
+                return 'Token not found';
+            }
         } else {
-            return 'No';
+            return 'No token provided';
         }
     }
 
     public function updatepass(Request $request)
     {
         $request->validate([
-            'password' => 'required|string|min:6|confirmed',
-            'confirm_pass' => 'required|string|min:6|same:password',
+            'password' => 'required|string|min:6',
+            'confirm_pass' => 'required|same:password',
         ]);
 
-        $user = User::find($request->id);
-        $user->password = Hash::make($request->password);
-        $user->save();
-
-        password_reset::where('email', $user->email)->delete();
-
+        $user = User::where('email',$request->email)->first();
+        // dd($request->email);
+        if($user)
+        {
+            $user->update([
+                'password' => Hash::make($request->password),
+            ]);
+            password_reset::where('email', $user->email)->delete();
+        }
         return "<h1>Your Password Reset was Successful!</h1>";
 
     }
